Adding inline documentation
